Save/Load Functionality Bugfixes
Updated Mark's code to fix the basic save/load functionality within SIRS. The program will now serialize the course and assignment objects and output them to a file in the main SIRS file structure called "courses.dat" when the "Save Data" button is pressed. Then, when the "Load Data" button is pressed, it will pull the information from that file and then display the information from the file in the display window.

There is a quite a lot more we can do with this, but for now it meets the objective requirement of a save/load system!

Changelog:
Course & Assignment classes:
- Implemented "Serializable" interface to enable the ObjectOutputStream to serialize them

DataPersistenceManager class:
- Renamed file to correct classname
- Added to proper Java package
- added close() method execution to ensure file saves properly
- Fixed minor syntax and style problems

Controller class:
- Added handling for "Save Data" and "Load Data" from getActionCommand()
- Removed redundant saveCourses() and loadCourses() functions
- loadCourses now sets the program's course object to the loaded one and updates the assigment UI display
- Added a thrown IllegalArgumentExpection if the course loaded is null
